Made it possible for admins and superadmins to set a vacation message, edit vacation messages, and remove them.
main.php added to admin directory so admin and / can share more
templates.

the new edit-vacation.php in admin and / are identical except
for template references

languages/en.lang and languages/default.lang are in sync now

I attempted to add some of the new .lang variables I created to
the spanish file (I don't know it that well, hope its OK)

fixed a bug in users/vacation.php

Added a "Main" menu item to the admin menu that gets you back to where
you started when you log in.